requirements.txt
Package    Version
---------- ---------
certifi    2020.6.20
chardet    3.0.4
ffmpeg     1.4
idna       2.10
lxml       4.5.2
Pillow     8.0.1
pip        20.2.3
pymongo    3.11.0
requests   2.24.0
selenium   3.141.0
setuptools 49.6.0
urllib3    1.25.10
wheel      0.35.1
